@@ -3,50 +3,37 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "serve": "vue-cli-service serve --host 0.0.0.0",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "dev": "vite --host 0.0.0.0",
+    "build": "vite build",
+    "serve": "vite preview",
+    "lint": "eslint --ext .js,.vue --ignore-path .gitignore --fix src"
   },
   "dependencies": {
-    "@babel/polyfill": "^7.11.5",
-    "bootstrap-icons": "^1.7.2",
-    "core-js": "^3.20.1",
-    "csv": "^6.0.5",
+    "csv-parse": "^5.3.0",
+    "csvtojson": "^2.0.10",
     "dayjs": "^1.10.7",
     "element-ui": "^2.15.9",
+    "events": "^3.3.0",
     "loaders.css": "^0.1.2",
     "mutationobserver-shim": "^0.3.7",
-    "portal-vue": "^2.1.7",
-    "svg-url-loader": "^7.1.1",
-    "vue": "^2.6.11",
-    "vue-virtual-scroll-list": "^2.3.3",
-    "webpack-bundle-analyzer": "^4.5.0"
+    "stream-browserify": "^3.0.0",
+    "vue": "^2.7.10",
+    "vue-virtual-scroll-list": "^2.3.3"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-eslint": "~4.5.0",
-    "@vue/cli-service": "~4.5.0",
-    "babel-eslint": "^10.1.0",
-    "babel-plugin-component": "^1.1.1",
-    "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^6.2.2",
+    "@esbuild-plugins/node-globals-polyfill": "^0.1.1",
+    "@vitejs/plugin-vue2": "^1.1.2",
+    "bootstrap-icons": "^1.9.1",
+    "eslint": "^8.23.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-vue": "^8.7.1",
+    "prettier": "^2.7.1",
     "sass": "^1.45.2",
-    "sass-loader": "^10.0.2",
-    "vue-template-compiler": "^2.6.11"
-  },
-  "eslintConfig": {
-    "root": true,
-    "env": {
-      "node": true
-    },
-    "extends": [
-      "plugin:vue/essential",
-      "eslint:recommended"
-    ],
-    "parserOptions": {
-      "parser": "babel-eslint"
-    },
-    "rules": {}
+    "svgo": "^2.8.0",
+    "unplugin-vue-components": "^0.22.4",
+    "vite": "^2.5.4",
+    "vite-plugin-compression": "^0.5.1"
   },
   "browserslist": [
     "> 1%",
