@@ -5,14 +5,36 @@
       <div class="title-detail">
         <div>来听听草原上小狼和小羊的歌声吧！</div>
         <div>录音棚收录了几乎全部呜米和咩栗直播唱的歌</div>
-        <div>还没关注
-          <a href="https://space.bilibili.com/617459493" target="_blank" rel="noreferrer noopener">@呜米 </a>
-          <a href="https://space.bilibili.com/745493" target="_blank" rel="noreferrer noopener">@咩栗</a>
+        <div>
+          还没关注
+          <a
+            href="https://space.bilibili.com/617459493"
+            target="_blank"
+            rel="noreferrer noopener"
+            >@呜米
+          </a>
+          <a
+            href="https://space.bilibili.com/745493"
+            target="_blank"
+            rel="noreferrer noopener"
+            >@咩栗</a
+          >
           ？搞快关注！
         </div>
-        <div>网易云音乐：
-          <a href="https://music.163.com/#/user/home?id=1306304320" target="_blank" rel="noreferrer noopener">@呜米嗷嗷嗷 </a>
-          <a href="https://music.163.com/#/user/home?id=1356709812" target="_blank" rel="noreferrer noopener">@咩栗咩咩咩</a>
+        <div>
+          网易云音乐：
+          <a
+            href="https://music.163.com/#/user/home?id=1306304320"
+            target="_blank"
+            rel="noreferrer noopener"
+            >@呜米嗷嗷嗷
+          </a>
+          <a
+            href="https://music.163.com/#/user/home?id=1356709812"
+            target="_blank"
+            rel="noreferrer noopener"
+            >@咩栗咩咩咩</a
+          >
         </div>
       </div>
     </div>
@@ -21,10 +43,9 @@
 </template>
 
 <script>
-
 export default {
-  name: 'Banner'
-}
+  name: "PageBanner",
+};
 </script>
 
 <style scoped>
