@@ -1,69 +1,42 @@
 <template>
   <div
-    v-bind:class="[
+    :class="[
       'c-playlist-song',
       { 'c-playlist-playing': index === current_song },
     ]"
-    v-on:click="$parent.$parent.$emit('apply', index)"
+    @click="$parent.$parent.$emit('apply', index)"
   >
     <div class="playlist-name">
       <span class="playlist-index">{{ index + 1 }}. </span>{{ source.name }}
     </div>
     <div class="playlist-dash">-</div>
-    <div class="playlist-artist">{{ source.artist }}</div>
-    <div class="playlist-status">{{ source.status }}</div>
-    <div class="playlist-duration">{{ source.duration }}</div>
+    <div class="playlist-artist">
+      {{ source.artist }}
+    </div>
+    <div class="playlist-status">
+      {{ source.status }}
+    </div>
+    <div class="playlist-duration">
+      {{ source.duration }}
+    </div>
     <div
       class="playlist-clear"
-      v-on:click.stop="$parent.$parent.$emit('remove', index)"
+      @click.stop="$parent.$parent.$emit('remove', index)"
     >
-      <div class="playlist-clear-img"></div>
+      <div class="playlist-clear-img" />
     </div>
   </div>
 </template>
 
-
-
-
-
-
-
-
-
-
-
 <script>
 export default {
   name: "PlayListItem",
+  props: ["index", "source", "current_song"],
   data() {
     return {};
   },
-  props: [
-    'index',
-    'source',
-    'current_song'
-  ]
 };
 </script>
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 
 <style scoped>
 .c-playlist-song {
@@ -134,7 +107,7 @@ export default {
 .playlist-clear-img {
   width: 1.5rem;
   height: 1.5rem;
-  background-image: url("~bootstrap-icons/icons/x.svg");
+  background-image: url("@/assets/ui/x.svg");
   background-position: center;
   background-size: contain;
 }
