@@ -1,18 +1,17 @@
 <template>
-  <pop-up-main
-    v-on:closepopup="$emit('closepopup')"
-    title="欢迎！"
-  >
+  <pop-up-main title="欢迎！" @closepopup="$emit('closepopup')">
     <div class="content">
       <p>看到你是首次打开，球球你关注一下在b站的MeUmy切歌man吧！！！</p>
-      <p>随便给你推荐几个：
+      <p>
+        随便给你推荐几个：
         <a
           v-for="c in cutter"
-          v-bind:key="c[0]"
-          v-bind:href="'https://space.bilibili.com/'+c[0]"
+          :key="c[0]"
+          :href="'https://space.bilibili.com/' + c[0]"
           target="_blank"
           rel="noreferrer noopener"
-        >@{{c[1]}}</a>
+          >@{{ c[1] }}</a
+        >
       </p>
       <p>快乐听歌鸭！~</p>
     </div>
@@ -20,23 +19,23 @@
 </template>
 
 <script>
-import PopUpMain from './Main.vue'
+import PopUpMain from "./Main.vue";
 
 export default {
-  name: 'PopUpInfo',
+  name: "PopUpInfo",
   components: {
-    PopUpMain
+    PopUpMain,
   },
   data() {
     return {
-      cutter: window.meumy.cutter_list
-    }
-  }
-}
+      cutter: window.meumy.cutter_list,
+    };
+  },
+};
 </script>
 
 <style scoped>
-.content >p >a {
+.content > p > a {
   margin-right: 1rem;
 }
 </style>
