@@ -1,35 +1,28 @@
 <template>
-  <div class="background" v-on:mousedown.self="$emit('closepopup')">
+  <div class="background" @mousedown.self="$emit('closepopup')">
     <div class="card c-popup">
       <div class="popup-title">
-        <div>{{title}}</div>
-        <div
-          class="popup-close"
-          v-on:click.stop="$emit('closepopup')"
-        >
-          <div class="popup-close-img"></div>
+        <div>{{ title }}</div>
+        <div class="popup-close" @click.stop="$emit('closepopup')">
+          <div class="popup-close-img" />
         </div>
       </div>
       <hr />
       <div class="popup-content">
-        <slot></slot>
+        <slot />
       </div>
     </div>
   </div>
 </template>
 
 <script>
-
 export default {
-  name: 'PopUpMain',
+  name: "PopUpMain",
+  props: ["title"],
   data() {
-    return {
-    }
+    return {};
   },
-  props: [
-    'title'
-  ]
-}
+};
 </script>
 
 <style scoped>
@@ -76,7 +69,7 @@ export default {
 .popup-close-img {
   width: 1.5rem;
   height: 1.5rem;
-  background-image: url("~bootstrap-icons/icons/x.svg");
+  background-image: url("@/assets/ui/x.svg");
   background-position: center;
   background-size: contain;
 }
