@@ -1,29 +1,19 @@
-import '@babel/polyfill'
-import 'mutationobserver-shim'
-import Vue from 'vue'
+import "mutationobserver-shim";
+import Vue from "vue";
 
-import { Select, Input, Option, Popover } from 'element-ui'
+import "@/js/global.js";
 
-import '@/js/global.js'
+import App from "./App.vue";
 
-import App from './App.vue'
+Vue.prototype.$ELEMENT = { size: "medium" };
 
-Vue.prototype.$ELEMENT = { size: 'medium' }
-Vue.use(Select)
-Vue.use(Input)
-Vue.use(Option)
-Vue.use(Popover)
-
-Vue.config.productionTip = false
-
-
-
+Vue.config.productionTip = false;
 
 const MeUmyPlayer = new Vue({
-  render: h => h(App),
-}).$mount('#app')
+  render: (h) => h(App),
+}).$mount("#app");
 
-window.MeUmyPlayer = MeUmyPlayer
+window.meumyPlayer = MeUmyPlayer;
 
 // 注册service worker
 /*
@@ -33,4 +23,3 @@ if ('serviceWorker' in navigator) {
   })
 }
 */
-
